Add support for HFN override
Support for HFN override added both at the driver
level and at the descriptor level.
A per-context value indicates if HFN override is enabled
on a certain context or not.
Two functions for processing packets are now available,
one for packets for which HFN override is not desired
and one for packets which need HFN value override.
One test-case has been added, based on mixed_descs_tests

Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>