NULL integrity + AES|SNOW encryption update
TS 133.401 indicates in Annex B, paragraph B.0 that the
implementation of EIA0 should generate a MAC of all zeros,
of length 4 bytes. This was not implemented in the previous
versions of the PDCP SEC driver, which assumed that no
MAC was generated for EIA0. This patch rectifies this problem.
Tests were updated to match the fixing (test_sec_driver,
mixed_descs_tests, hfn_override_tests)

Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>