Fix contexts not properly freed if packets are in flight
If packets are in flight, the context isn't outright destroyed, but
put in a temporary list and only when the last packet has been processed
the context is free'ed and then ready to be reused. But in this case
the producer and consumer indexes are not reset to 0, as in the case
where the context is deleted when there are no packets in flight.
When get_free_context() is invoked, and the previously free'ed
context with packets in flight is retrieved, then an assertion will
fail, because the producer and consumer indexes are not 0.

Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>