Add proper SNOW f8/AES CMAC decapsulation support
PDCP decapsulation requires the ICV check to be done after the message
has decrypted. Due to SNOW f8 and AES being both CLASS 1 EUs, this means that
the process cannot happen in parallel (through in or outsnooping). Another issue
is the length of the descriptor, which cannot exceed 54 WORDs for 32 bit pointers
and 51 WORDs for 36 bit pointers. In order to reduce the descriptor word count,
non-used parts from PDB were discarded and overwriting of the JD is employed.

Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>