Fix mutex not taken if non DEBUG mode
ASSERT macro is defined only if DEBUG is defined, otherwise
the expression under it is not evaluated. Thus, for lists
which need mutexes to be aquired, the operation would take
place without any protection. This patch adds the necessary
glue code that enables the mutex_lock functions to be executed
in either configurations.

Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>