sec-driver, unit & system TCs: updates for 64 bit processing
Initially, the USPDCP driver was designed to work on 32bit only
platforms (BSC 9131 & BSC 9132). Once the porting to DPAA
enabled platforms was done, the need for 64 bit support
has arisen. This patch adds the necessary glue code for enabling
the driver to work on both 32bit & 64 bit kernel/userspace
ombinations.

Change-Id: Ia1cf3e26dde191350ba8d3e50b8d7e94fcd1193f
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/11830