test_sec_driver_b2b, test_sec_driver_(single_th)_sg: fix SG fragment size
The calculation for the SG fragment size is not done properly, by ignoring
the boundary cases (0 fragments and MAX_SG_TBL_ENTRIES) cases. Plus, the
length of the SG fragment is improperly calculated. This patch solves
these issues by employing a repeated substraction method for the
calculus of the SG fragments' sizes.

Change-Id: I482657c6ac5d544707d87619ac1b0f12041fc5e4
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/11831